Migrate legacy CLI stuff to BriefcaseCLI - This will let us simplify the GUI launcher class - This also improves separation of concerns
